Moved director only routes to director.py - Fixed test_admin and test_director accordingly
