+++ apps/api/src/routers/director.py
@@ -1,18 +1,28 @@
+import asyncio
+
 from datetime import datetime
 from logging import getLogger
-from typing import Annotated, Any, Optional
+from typing import Annotated, Any, Literal, Optional, Sequence
 
 from fastapi import APIRouter, Body, Depends, HTTPException, status
 from pydantic import BaseModel, EmailStr, TypeAdapter, ValidationError
 
+from admin import applicant_review_processor
 from auth.authorization import require_role
 from auth.user_identity import User, uci_email, utc_now
-from models.user_record import Role
+from models.ApplicationData import Decision
+from models.user_record import Role, Status
 from services import mongodb_handler, sendgrid_handler
 from services.mongodb_handler import BaseRecord, Collection
-from services.sendgrid_handler import PersonalizationData, Template
-from routers.admin import recover_email_from_uid
+from services.sendgrid_handler import (
+    ApplicationUpdatePersonalization,
+    PersonalizationData,
+    Template,
+)
+from routers.admin import recover_email_from_uid, retrieve_thresholds
+from utils import email_handler
 from utils.email_handler import IH_SENDER
+from utils.batched import batched
 
 log = getLogger(__name__)
 
@@ -209,3 +219,226 @@ async def apply_reminder(user: Annotated[User, Depends(require_director)]) -> No
             personalizations,
             True,
         )
+
+
+@router.post("/rsvp-reminder", dependencies=[Depends(require_director)])
+async def rsvp_reminder() -> None:
+    """Send email to applicants who have a status of ACCEPTED or WAIVER_SIGNED
+    reminding them to RSVP."""
+    # TODO: Consider using Pydantic model validation instead of type annotations
+    not_yet_rsvpd: list[dict[str, Any]] = await mongodb_handler.retrieve(
+        Collection.USERS,
+        {
+            "roles": Role.APPLICANT,
+            "status": {"$in": [Decision.ACCEPTED, Status.WAIVER_SIGNED]},
+        },
+        ["_id", "first_name"],
+    )
+
+    personalizations = []
+    for record in not_yet_rsvpd:
+        personalizations.append(
+            ApplicationUpdatePersonalization(
+                email=recover_email_from_uid(record["_id"]),
+                first_name=record["first_name"],
+            )
+        )
+
+    log.info(f"Sending RSVP reminder emails to {len(not_yet_rsvpd)} applicants")
+
+    await sendgrid_handler.send_email(
+        Template.RSVP_REMINDER,
+        IH_SENDER,
+        personalizations,
+        True,
+    )
+
+
+@router.post("/confirm-attendance", dependencies=[Depends(require_director)])
+async def confirm_attendance() -> None:
+    """Update applicant status to void or attending based on their current status."""
+    # TODO: consider using Pydantic model, maybe BareApplicant
+    records = await mongodb_handler.retrieve(
+        Collection.USERS, {"roles": Role.APPLICANT}, ["_id", "status"]
+    )
+
+    statuses = {
+        Status.CONFIRMED: Status.ATTENDING,
+        Decision.ACCEPTED: Status.VOID,
+        Status.WAIVER_SIGNED: Status.VOID,
+    }
+
+    for status_from, status_to in statuses.items():
+        curr_records = [record for record in records if record["status"] == status_from]
+
+        for record in curr_records:
+            record["status"] = status_to
+
+        log.info(
+            f"Changing status of {len(curr_records)} from {status_from} to {status_to}"
+        )
+
+        await asyncio.gather(
+            *(
+                _process_status(batch, status_to)
+                for batch in batched(
+                    [str(record["_id"]) for record in curr_records], 100
+                )
+            )
+        )
+
+
+@router.post("/set-thresholds")
+async def set_hacker_score_thresholds(
+    user: Annotated[User, Depends(require_director)],
+    accept: float = Body(),
+    waitlist: float = Body(),
+) -> None:
+    """
+    Sets accepted and waitlisted score thresholds.
+    Any score under waitlisted is considered rejected.
+    """
+
+    thresholds: Optional[dict[str, float]] = await retrieve_thresholds()
+
+    if accept != -1 and thresholds is not None:
+        thresholds["accept"] = accept
+    if waitlist != -1 and thresholds is not None:
+        thresholds["waitlist"] = waitlist
+
+    if (
+        accept < -1
+        or accept > 10
+        or waitlist < -1
+        or waitlist > 10
+        or (accept != -1 and waitlist != -1 and waitlist > accept)
+        or (thresholds and thresholds["waitlist"] > thresholds["accept"])
+    ):
+        log.error("Invalid threshold score submitted.")
+        raise HTTPException(status.HTTP_400_BAD_REQUEST)
+
+    log.info("%s changed thresholds: Accept-%f | Waitlist-%f", user, accept, waitlist)
+
+    # negative numbers should not be received, but -1 in this case
+    # means there is no update to the respective threshold
+    update_query = {}
+    if accept != -1:
+        update_query["accept"] = accept
+    if waitlist != -1:
+        update_query["waitlist"] = waitlist
+
+    try:
+        await mongodb_handler.raw_update_one(
+            Collection.SETTINGS,
+            {"_id": "hacker_score_thresholds"},
+            {"$set": update_query},
+            upsert=True,
+        )
+    except RuntimeError:
+        log.error(
+            "%s could not change thresholds: Accept-%f | Waitlist-%f",
+            user,
+            accept,
+            waitlist,
+        )
+        raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+
+@router.post("/release/mentor-volunteer", dependencies=[Depends(require_director)])
+async def release_mentor_volunteer_decisions() -> None:
+    """Update applicant status based on decision and send decision emails."""
+    mentor_records = await mongodb_handler.retrieve(
+        Collection.USERS,
+        {"status": Status.REVIEWED, "roles": {"$in": [Role.MENTOR]}},
+        ["_id", "application_data.reviews", "first_name"],
+    )
+
+    for record in mentor_records:
+        applicant_review_processor.include_review_decision(record)
+
+    volunteer_records = await mongodb_handler.retrieve(
+        Collection.USERS,
+        {"status": Status.REVIEWED, "roles": {"$in": [Role.VOLUNTEER]}},
+        ["_id", "application_data.reviews", "first_name"],
+    )
+
+    for record in volunteer_records:
+        applicant_review_processor.include_review_decision(record)
+
+    await _process_records_in_batches(mentor_records, Role.MENTOR)
+    await _process_records_in_batches(volunteer_records, Role.VOLUNTEER)
+
+
+@router.post("/release/hackers", dependencies=[Depends(require_director)])
+async def release_hacker_decisions() -> None:
+    """Update hacker applicant status based on decision and send decision emails."""
+    records = await mongodb_handler.retrieve(
+        Collection.USERS,
+        {"status": Status.REVIEWED, "roles": {"$in": [Role.HACKER]}},
+        ["_id", "application_data.reviews", "first_name"],
+    )
+
+    thresholds: Optional[dict[str, float]] = await retrieve_thresholds()
+
+    if not thresholds:
+        log.error("Could not retrieve thresholds")
+        raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+    for record in records:
+        applicant_review_processor.include_hacker_app_fields(
+            record, thresholds["accept"], thresholds["waitlist"]
+        )
+
+    await _process_records_in_batches(records, Role.HACKER)
+
+
+async def _process_status(uids: Sequence[str], status: Status) -> None:
+    ok = await mongodb_handler.update(
+        Collection.USERS, {"_id": {"$in": uids}}, {"status": status}
+    )
+    if not ok:
+        raise RuntimeError("gg wp")
+
+
+async def _process_records_in_batches(
+    records: list[dict[str, object]],
+    application_type: Literal[Role.HACKER, Role.MENTOR, Role.VOLUNTEER],
+) -> None:
+    for decision in (Decision.ACCEPTED, Decision.WAITLISTED, Decision.REJECTED):
+        group = [record for record in records if record["decision"] == decision]
+        if not group:
+            continue
+        await asyncio.gather(
+            *(
+                _process_batch(batch, decision, application_type)
+                for batch in batched(group, 100)
+            )
+        )
+
+
+async def _process_batch(
+    batch: tuple[dict[str, Any], ...],
+    decision: Decision,
+    application_type: Literal[Role.HACKER, Role.MENTOR, Role.VOLUNTEER],
+) -> None:
+    uids: list[str] = [record["_id"] for record in batch]
+    log.info(f"Setting {application_type}s {','.join(uids)} as {decision}")
+    ok = await mongodb_handler.update(
+        Collection.USERS, {"_id": {"$in": uids}}, {"status": decision}
+    )
+    if not ok:
+        raise RuntimeError("gg wp")
+
+    # Send emails
+    log.info(
+        f"Sending {application_type} {decision} emails for {len(batch)} applicants"
+    )
+    await email_handler.send_decision_email(
+        map(_extract_personalizations, batch), decision, application_type
+    )
+
+
+def _extract_personalizations(decision_data: dict[str, Any]) -> tuple[str, EmailStr]:
+    name = decision_data["first_name"]
+    email = recover_email_from_uid(decision_data["_id"])
+    return name, email
